+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+Improve the repository architecture. Reader class now takes care of raw
+file interaction (image and metadata).
+
+## Changed
+
+- Reader class now takes care of raw file interaction (image and metadata).
+
 ## [0.2.1]
 
 ### Added